fix some bugs

+++ src/Professor.java
@@ -1,7 +1,7 @@
 public class Professor extends person {
     private int jalase;
-    private double hogog;
-    public Professor (String firstname, String lastname, int jalase, double hogog) {
+    private long hogog;
+    public Professor (String firstname, String lastname, int jalase, long hogog) {
         super(firstname, lastname);
         this.jalase = jalase;
         this.hogog = hogog;
@@ -12,7 +12,7 @@ public class Professor extends person {
     public double gethogog () {
         return hogog;
     }
-    public double calculateSalary () {
+    public long calculateSalary () {
         return this.hogog * this.jalase;
     }
 }
